feat: [001-SIGN-UP] 회원 interface, service 통합
단위테스트코드 작성이 어렵고, 분리할 이유가 없어서 통합.
