+++ study/default-study.md
@@ -1,0 +1,21 @@
+# MY E-COMMERCE PROJECT STUDY
+2025 간단한 이커머스 서비스를 만들면서 공부한 내용
+
+---
+### < service, service impl 분리하지 않는 이유 >
+- 참고블로그 https://zhfvkq.tistory.com/76
+
+spring은 객체간의 의존성을 관리해주는데 이 의존성은 테스트 코드 단위테스트를 작성할때 문제가 생긴다.    
+의존성을 가지는 다른 객체에 의해 테스트 결과가 영향을 받는다.    
+그래서 내가 원하는 동작만 하도록 하는것이 Mock 객체다.    
+이를 편하게 사용하도록 지원하는것이 Mockito 이다.    
+
+Service와 ServiceImpl 구조를 사용하면 단위 테스트가 어려울 수 있다.    
+이 구조를 사용하면 Service 인터페이스와 ServiceImpl 구현체를 모두 생성해야 하므로 테스트 코드를 작성하기 어렵다.    
+또한, 구현체와 인터페이스 간의 의존성이 높아져 테스트 케이스 작성이 복잡해질 수 있다.     
+결론적으로 Service ServiceImpl 구조를 사용해야 할까?    
+OOP 관점에서 봤을 때 인터페이스는 다형성 혹은 개방 폐쇄 원칙 때문에 사용한다. 보통 흔히 얘기하는 느슨한 결합 혹은 유연해지도록 설계하기 위해 사용된다.    
+그렇다면 Spring에서 Service는 다형성이 필요한가?    
+우리가 사용하는 대부분의 Service는 하나의 Service가 여러 개의 SerivceImpl을 가지고 있는 경우가 아닌 1:1 구조를 띄고 있다.    
+그러므로 Service를 사용하고, 추후 다형성이 필요해지면 그때 해당 Service에 대해서 인터페이스를 적용해도 늦지 않는 것 같다.    
+
